requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
asttokens==2.4.1
astunparse==1.6.3
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.1
confection==0.1.4
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.1
decorator==5.1.1
dm-tree==0.1.8
executing==2.0.1
flatbuffers==23.5.26
fonttools==4.49.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.0
h5py==3.10.0
idna==3.6
ipykernel==6.29.3
ipython==8.22.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==3.0.5
keras-nightly==3.1.0.dev2024022803
keras-nlp==0.0.2
kiwisolver==1.4.5
langcodes==3.3.0
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
murmurhash==1.0.10
namex==0.0.7
nest-asyncio==1.6.0
numpy==1.26.4
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
plotly==5.19.0
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
pywin32==306
pyzmq==25.1.2
requests==2.31.0
rich==13.7.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
smart-open==6.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
tb-nightly==2.17.0a20240228
tenacity==8.2.3
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tf-nightly==2.17.0.dev20240227
tf_nightly_intel==2.17.0.dev20240227
thinc==8.2.3
threadpoolctl==3.3.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
Werkzeug==3.0.1
wrapt==1.16.0
